phase distortion runs in 32 bit now

+++ dsp/phase_distorter.hpp
@@ -21,8 +21,8 @@ class PhaseDistorter {
         uint16_t* warp_amt_;
         float blend_amt_;
         float s1, s2, o2, kink_amt;
-        uint16_t kink_point;
-        uint16_t kink_value(uint16_t input_phase);
+        uint32_t kink_point;
+        uint32_t kink_value(uint32_t input_phase);
 };
